rust refactor complete
